requirements.txt
django
django-extensions
django-grappelli
sqlparse
django-crispy-forms
